feat: add route /c-by-id/<projectId>/update and redirect /projects-by-id/<projectId>/update
