+++ pages/projects-by-id/[projectId]/update.tsx
@@ -1,22 +1,20 @@
 import { GetServerSideProps } from "next";
 
-import PageUpdateProjectV2 from "../../../containers/PageUpdateProjectV2";
-
-type Props = {
-  projectId: string;
-};
-
-export default function RouteToPageUpdateProjectV2({ projectId }: Props) {
-  return <PageUpdateProjectV2 projectId={projectId} />;
+export default function RouteToPageUpdateProjectV2() {
+  return null;
 }
 
-export const getServerSideProps: GetServerSideProps<Props> = async (
-  context
-) => {
-  const { projectId } = context.params || {};
+export const getServerSideProps: GetServerSideProps = async (context) => {
+  const projectId = context.query["projectId"];
+
   if (typeof projectId !== "string" || !/^[0-9A-Fa-f]+$/.test(projectId)) {
     return { notFound: true };
   }
 
-  return { props: { projectId } };
+  return {
+    redirect: {
+      permanent: true,
+      destination: `/c-by-id/${projectId}/update`,
+    },
+  };
 };
